Tweak order of link parsing
