+++ gradle.properties
@@ -11,7 +11,7 @@ loader_version=0.14.8
 fabric_version=0.59.0+1.19.2
 
 # Mod Properties
-	mod_version = 2.0.1+1.19.2
+	mod_version = 2.0.2+1.19.2
 	maven_group = eu.pb4
 	archives_base_name = styled-chat
